[libc] Fix wrappergen_test.cpp on runtimes build
The `tools` directory depends on `llvm_gtest` which is not immediately
availible with a runtimes build. This patch builds the `llvm_gtest`
target if it isn't present. Additionally, we use the CMake binary
directory to find the tool binary, which is different when using a
runtimes build. Using the LLVM binary directory should match both build
conditions.

Reviewed By: lntue

Differential Revision: https://reviews.llvm.org/D139348

GitOrigin-RevId: 1dfb0cd703fe1ada21681283fd75e6aba4b7f216